requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
decorator==5.1.1
deta==1.2.0
dnspython==2.4.2
duckdb==0.9.1
extra-streamlit-components==0.1.60
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.37
google-api-core==2.12.0
google-api-python-client==2.103.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
gsheetsdb==0.1.13.1
gspread==5.11.3
gspread-dataframe==3.3.1
gspread-formatting==1.1.2
gspread-pandas==3.2.3
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mo-dots==4.22.21108
mo-future==3.147.20327
mo-imports==3.149.20327
mo-kwargs==4.22.21108
mo-logs==4.23.21108
moz-sql-parser==4.40.21126
multidict==6.0.4
numpy==1.26.0
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.28.1
packaging==23.2
pandas==1.5.3
Pillow==10.0.1
plotly==5.17.0
protobuf==4.24.4
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
pymongo==4.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-docx==1.0.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
six==1.16.0
smmap==5.0.1
sql-metadata==2.9.0
sqlparse==0.4.4
st-gsheets-connection==0.0.3
streamlit==1.27.2
streamlit-aggrid==0.3.4.post3
streamlit-authenticator==0.2.3
streamlit-option-menu==0.3.6
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
typing-extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
uritemplate==4.1.1
urllib3==2.0.6
validators==0.22.0
watchdog==3.0.0
yarl==1.9.2
zipp==3.17.0
